+++ src/comments.js
@@ -3,10 +3,10 @@ var types = require("ast-types");
 var n = types.namedTypes;
 var isArray = types.builtInTypes.array;
 var isObject = types.builtInTypes.object;
-var pp = require("./pp");
-var fromString = pp.fromString;
-var concat = pp.concat;
-var hardline = pp.hardline;
+var docBuilders = require("./doc-builders");
+var fromString = docBuilders.fromString;
+var concat = docBuilders.concat;
+var hardline = docBuilders.hardline;
 var util = require("./util");
 var comparePos = util.comparePos;
 var childNodesCacheKey = require("private").makeUniqueKey();
